Refactor AI send queue
This heavily simplifies the logic, and allows us to skip straight to the next socket ID to attempt.
The official logic essentially just assumes there's a maximum of 10 connected sockets out of the pool of 100,
it doesn't really handle multiple servers.

Now we will consistently attempt to cycle through them all once, starting from the next socket ID to send to.