Remove StringConversion helpers, move DebugUtils to N3Base
StringConversion helpers were used by CIni (which no longer bothers to use wide-char variants, so translation is unnecessary),
and various path fetches in the server.

This was looked at here because of the path translations; they were inconsistent, not always re-encoding appropriately.
Now they're all consistent and this has been removed.

StringConversion also depended on Windows (MultiByteToWideChar/WideCharToMultiByte calls), so this is a good one to remove.

Also moved DebugUtils because the server uses spdlog::trace() instead; only the client & its tools use TRACE().
In this way we remove the other hard Windows.h dependency in the server while we're at it.