Server: Use the console sink by default
As OnStart() is what sets up the loggers, and rendering occurs afterwards,
we have the ftxui sink just use the console logger to log on-demand as it's loading
up until it's finished loading and it sets up the ftxui sink for rendering.

We rename _useStdout to _useConsoleSink to be more explicit, and add _storeLogBuffer
as everything we're writing to console now should be swapped over to & rendered by ftxui.