MSBuild: Guard git access
Parallel builds can unintentionally use git at the same time.
This causes it to hit git's internal lock, which occasionally fails builds. A rebuild usually fixes this, unless it was unlucky enough to corrupt the git index.

MKDIR is atomic, so we basically just guard it behind a created directory.
The only annoying part is the delay; the ping trick is the only thing that really seems to behave in this environment. We could run it through Powershell but
invoking that is slower, so I'd rather not.