Server: Word wrap log entries as best we can
Ideally, paragraph needs to support colourisation.
Since it doesn't, we're forced to use multiple text() calls, which it wraps on individually.

It also groups them horizontally, so it will only start wrapping from the point the previous elements end at.

It's *very* simple with just paragraph(entireLogLine), but unfortunately we can't use colours that way.

That being the case, we tweak the console log default to colourise "[info] " so it "nicely" wraps after that.

Ideally I'd just have it wrap properly to the start of the next line, but short of writing our own element to
handle this paragraph behaviour, I unfortunately don't think this is actually possible.