Change default startup overrides for solutions
Additionally, add new *.slnLaunch files for our multiple startup projects.
Note that there's no way of actually setting it to a profile, it will always
default to the first entry in the solution (unless you've specifically told it otherwise).

As such, this means that generally speaking, we no longer default to say, the shared project
anymore, but it will always default to a singular startup project.

You can check the startup configurations to swap to 'multiple startup projects', where you'll
see our existing profiles and be able to select them.

Resolves #668