Move math code to MathUtils & share it with the server
Move all of the client's math code to its own independent MathUtils project.
Since it's no longer tied to D3DX, we can share all of this with the server, replacing the code in both with the singular shared implementation.

MathUtils-tests now replaces the existing client-tests (as there's no other client tests to speak of right now), which can also now be run by CMake,
and on compilers other than MSVC.