Replace audio engine with OpenAL (#719)
We replace the DirectSound implementation with OpenAL (specifically OpenAL-soft).

Previously, we had tacked on mpg123 to pre-decode the MP3s to WAVs (which were stored to file for future lookups), however this has been replaced with on-demand live decoding of chunks, so nothing needs to be stored (which also means we no longer need to find a suitable OS-specific path. Yay!)

From an outside perspective, the API is much the same, but I removed a bunch of public methods which weren't really used by anything or were otherwise pointless.

See #719 for more information.