Strip DX9 SDK implementation so we can simplify include paths (#723)
The DXSDK contains many files which clash with the Windows SDK. As we can't necessarily control include order (without the hacky property sheet approach), we'll (unfortunately) just strip our implementation to something close to "bare minimum" (being generous - there's still a lot of excess there).

It contained DX10, DX11, etc, which has been stripped down to just DX9 (with a couple of exceptions to be safe), to avoid collisions with the Windows SDK (which tends to pop up within the MFC tools).

With the property sheet no longer needed, we don't necessarily require the VS generator anymore, though this isn't strictly tested/supported.