Ebenezer: Remove redundant CUser::RunNpcEvent() (#727)
This is basically just CUser::RunEvent() with the NPC pointer passed in.
RunEvent() also handles the outer loop, where RunNpcEvent() is expected to be run
in a loop.

Other than that, the only real difference is RunNpcEvent() doesn't handle EXEC_SAVE_COM_EVENT
but there's no real functional reason for that.

Regardless, RunNpcEvent() isn't even a thing in the official 1.298 binaries; they're shared there.